+++ packages/plugin-inquirer-cli/src/Helpers.types.ts
@@ -1,6 +1,5 @@
 /*
-Copyright 2019 Javier Brea
-Copyright 2019 XbyOrange
+Copyright 2023 Javier Brea
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -9,8 +8,11 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-"use strict";
+/** Status of header. It defines color to be used */
+export type HeaderStatusLevel = 0 | 1 | 2;
 
-const InquirerCli = require("./src/Cli");
+/** Header color */
+export type HeaderColor = "green" | "yellow" | "red";
 
-module.exports = InquirerCli;
+/** Message to display in header */
+export type HeaderMessage = number | string | boolean;