+++ packages/plugin-inquirer-cli/src/index.ts
@@ -9,15 +9,9 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const index = require("../src/index");
-const { Plugin } = require("../src/Plugin");
+import { Plugin } from "./Plugin";
 
-describe("index", () => {
-  it("should export the Cli plugin constructor", () => {
-    expect(index.Plugin).toEqual(Plugin);
-  });
+export * from "./types";
+export { Plugin } from "./Plugin";
 
-  it("should export the Cli plugin constructor by default", () => {
-    expect(index.default).toEqual(Plugin);
-  });
-});
+export default Plugin;