+++ packages/plugin-inquirer-cli/test/src/Inquirer.mocks.js
@@ -1,5 +1,6 @@
+/* eslint-disable */
 /*
-Copyright 2019-2022 Javier Brea
+Copyright 2019-2023 Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
@@ -11,7 +12,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 const sinon = require("sinon");
 
-const inquirer = require("../../src/Inquirer");
+const { Inquirer } = require("../../src/Inquirer");
 
 class CallBackRunner {
   constructor() {