+++ packages/plugin-inquirer-cli/test/src/Inquirer.spec.js
@@ -13,7 +13,7 @@ const inquirer = require("inquirer");
 const sinon = require("sinon");
 const chalk = require("chalk");
 
-const Inquirer = require("../../src/Inquirer").Inquirer;
+const { Inquirer } = require("../../src/Inquirer");
 
 const fooQuestions = {
   main: {