+++ test/main-e2e/src/createServer.spec.js
@@ -334,7 +334,7 @@ describe("createServer", () => {
       const response = await doFetch("/api/mock/custom-route-variants", {
         port: 3110,
       });
-      expect(response.body).toEqual(["get-user:real"]);
+      expect(response.body).toEqual([{ id: "get-user:real" }]);
     });
 
     it("should serve user 1 under the /api/users/1 path", async () => {