+++ packages/config/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
+## TODO
+- Callback type depending on option type
+
 ## [Unreleased]
 
 ### Changed